fix(sim): round spawn, pellet, damage-pool and typo math per x87 op
The original runs gameplay at 24-bit x87 precision, so every add, multiply
and divide rounds to float32. These paths computed in doubles and rounded
once, drifting by an ulp; both ports now follow the native operation order,
float literals and store points:

- creature_spawn_template: size/health/speed/reward/contact formulas, random
  fields, AI7, hardcore and retry scaling, spawn slot intervals, template
  tint literals, stale 0x0e root health, 0x02 fallback
- pellet jitter and speed scale in player_update
- projectile damage pool subtractions in projectile_update
- rush_mode_update spawn position and speed
- Typ-o spawn tint, position, heading, size and speed
- zig: revert the f64 mirroring introduced to match the old python math

+++ crimson-zig/src/runtime/fire_recipes.zig
@@ -7,28 +7,26 @@ const weapon_data = @import("weapon_data.zig");
 pub const WeaponId = game_ids.WeaponId;
 pub const ProjectileTypeId = game_ids.ProjectileTypeId;
 
-/// Pellet jitter and speed-scale rules are evaluated in double precision and
-/// rounded once when stored in the f32 projectile fields.
 pub const PelletJitterRule = union(enum) {
     none,
     modulo_centered: struct {
         modulo: u32,
         center: i32,
-        step: f64,
+        step: f32,
     },
     mask_centered: struct {
         mask: u32,
         center: i32,
-        step: f64,
+        step: f32,
     },
 };
 
 pub const SpeedScaleRule = union(enum) {
     none,
     modulo: struct {
-        base: f64,
+        base: f32,
         modulo: u32,
-        step: f64,
+        step: f32,
     },
 };
 
@@ -91,7 +89,7 @@ const gauss_ion_speed_scale: SpeedScaleRule = .{
     },
 };
 
-pub fn pelletJitterStepForWeapon(weapon_id: WeaponId) f64 {
+pub fn pelletJitterStepForWeapon(weapon_id: WeaponId) f32 {
     return switch (weapon_id) {
         .shotgun, .jackhammer => 0.0013,
         .sawed_off_shotgun => 0.004,
